+++ src/settings/MatchProfiles/MatchProfiles.js
@@ -154,6 +154,10 @@ export class MatchProfiles extends Component {
   defaultNewRecordInitialValues = {
     name: '',
     description: '',
+    /* TODO: these values are hardcoded now and will need to be changed in future (https://issues.folio.org/browse/UIDATIMP-175) */
+    incomingRecordType: 'MARC',
+    existingRecordType: 'HOLDINGS',
+    incomingDataValueType: 'STATIC_VALUE',
   };
 
   renderActionMenu = menu => {