+++ test/bigtest/network/scenarios/fetch-match-profiles-success.js
@@ -129,4 +129,10 @@ export default server => {
   server.get('/data-import-profiles/matchProfiles/:id');
 
   server.get('/data-import-profiles/profileAssociations/:id/masters', associatedJobProfiles);
+  server.post('/data-import-profiles/matchProfiles', (_, request) => {
+    const params = JSON.parse(request.requestBody);
+    const record = server.create('match-profile', params);
+
+    return record.attrs;
+  });
 };